requirements.txt
fastapi
fastapi-mail==1.3
uvicorn[standard]
pydantic==1.10.7
pydantic[email]
motor
python-dotenv
passlib[bcrypt]
pymongo
pymongo[srv]
python-multipart
python-jose[cryptography]
secrets